:art: move publish.py to register package
